Propagate XDG_CACHE_HOME to lit tests
This is another tool we can leverage to configure separate caches for CI
runs in Linux to hopefully reduce the likelihood of module cache
related issues.
Also this will allow us to respect the default values for the
environment we are running in.

As per clang source code, the environment variable only impacts Linux
environments.

This mimicks https://github.com/apple/swift/pull/35588

Addresses rdar://73833137