Fix SwiftDocC integration test
The integration test as currently written fails to actually compile
any documentation because the Symbol Graph files for the test Swift Package
are never built.

This updates the test to build Symbol Graph files for the test package,
convert it, and assert that produced DocC documentation contains the expected
`data` subdirectory.

Resolves rdar://91638258.